+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 * more time-related calls will accept time strings
 * added optional `delay` argument to `call_in_separate_thread`
 * beefed up BogusTerminableThread docs
-* added `ThreadCollection.add`
+* added functionality to `ThreadCollection`